# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "visionatrix-client",
    "version": "0.8.0.dev0"
  },
  "paths": {
...
    "/api/setting": {
      "get": {
        "summary": "Setting Get",
        "description": "Returns the value as a string or an empty string if the setting is not found.\n\nDefault endpoint for retrieving settings.\nUser settings have higher priority than global settings.",
        "operationId": "setting_get_api_setting_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The key of the setting to retrieve",
              "title": "Key"
            },
            "description": "The key of the setting to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved setting",
            "content": {
              "text/plain": {
                "example": "value_of_setting"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/global_setting": {
      "get": {
        "summary": "Global Setting Get",
        "description": "Retrieve the global setting value or an empty string if the global setting is not found.",
        "operationId": "global_setting_get_api_global_setting_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The key of the global setting to retrieve",
              "title": "Key"
            },
            "description": "The key of the global setting to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved global setting",
            "content": {
              "text/plain": {
                "example": "value_of_global_setting"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Global Setting Set",
        "description": "Creates, updates, or deletes a global setting.\n\nTo delete a setting, specify an empty string as the value.\nAccess is restricted to administrators only.",
        "operationId": "global_setting_set_api_global_setting_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_global_setting_set_api_global_setting_post"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Global setting updated successfully"
          },
          "401": {
            "description": "Unauthorized - Admin privilege required",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Admin privilege required"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user_setting": {
      "get": {
        "summary": "User Setting Get",
        "description": "Retrieve the user setting value or an empty string if the user setting is not found.",
        "operationId": "user_setting_get_api_user_setting_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The key of the user setting to retrieve",
              "title": "Key"
            },
            "description": "The key of the user setting to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user setting",
            "content": {
              "text/plain": {
                "example": "value_of_user_setting"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "User Setting Set",
        "description": "Creates, updates, or deletes a user setting.\n\nTo delete a setting, specify an empty string as the value.",
        "operationId": "user_setting_set_api_user_setting_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_user_setting_set_api_user_setting_post"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "User setting updated successfully"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_global_setting_set_api_global_setting_post": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "The key of the setting to update"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "The value of the setting to update, or an empty string to delete the setting"
          },
          "sensitive": {
            "type": "boolean",
            "title": "Sensitive",
            "description": "Flag that determines whether the value can be available to users or not"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "sensitive"
        ],
        "title": "Body_global_setting_set_api_global_setting_post"
      },
...
      "Body_user_setting_set_api_user_setting_post": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "The key of the setting to update"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "The value of the setting to update, or an empty string to delete the setting"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "Body_user_setting_set_api_user_setting_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}